==> Generate_figures_from_tables/4-Survival_CoxPH_multiVar_analyses.rmd <==
---
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## R Markdown
This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see <http://rmarkdown.rstudio.com>.

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

Load libraries
```{r echo=F}
library(dplyr)
library(tidyr)
library(ggplot2)
library(readxl)
```

Input tables
```{r}
combined_table_all <- readRDS(file="Combined_MAPKi_ICI_gene_gsva_tcrbcr_MCPcounter.rds")

#only the bulk RNAseq
ICI_datasets <- c("2017_Cell_NR","2020_NC_GA")
MAPKi_datasets <- c("2015_Cell_WH","2015_JCI_LK")

combined_table <- combined_table_all %>% 
                       filter(bioproject_name %in% c(ICI_datasets,MAPKi_datasets)) %>%
                       as.data.frame()

itemList <- colnames(combined_table)[17:ncol(combined_table)]

#remove variables with "proportions"
itemList0<- grep("proportion",itemList,value=T,invert=T)

#remove CXCR4, CD22, MCP counter of other immune cells from
# list of variable to test
#7/5/22 remove Chemokine gene Messina, germline IGH clonality and diversity
to_remove <- c("CXCR4","CXCR6","TGFB1","TCF7","BCL6","CD274","CCR7","IL21",
               "TOX","IL21R","CCL19","CD22","CCL21","IGHA1","IGHG1","IGHM","IGHD",
               "c2_cgp--DER_IFN_ALPHA_RESPONSE_UP",        
               "c2_cgp--RADAEVA_RESPONSE_TO_IFNA1_UP",     
               "c2_cgp--DER_IFN_BETA_RESPONSE_UP",
               "c2_cgp--MOSERLE_IFNA_RESPONSE",
               "h_all--HALLMARK_INTERFERON_ALPHA_RESPONSE",
               "x_all--CHEMOKINE_12_GENE_Messina",
               "SHM_occurrence_frequency",
               "Cytotoxic lymphocytes","Fibroblasts","Endothelial cells",
               "Neutrophils","Monocytic lineage","Monocytic lineage",
               "Myeloid dendritic cells","T cells","NK cells")

itemList1 <- setdiff(itemList0,to_remove)

#remove the prefix from the gene set names i.e. "c2_cgp--, h_all--" 
nameList1 <- gsub("^[\\w\\_]+\\-\\-","",itemList1,perl=T)
nameList1

name_mapping <- list()
for(i in 1:length(itemList1)){
  name_mapping[[itemList1[i]]] <- nameList1[i]
}
```

```{r}
library(survival)
library(survminer)
library(tidyverse)

generateKaplanMeier <- function(dataTable,varName,survival_label,draw.conf.int=T){
  fit <- survfit(Surv(Survival,Status)~Value,data = dataTable)
  y <- ggsurvplot(fit, data = dataTable,
           xlab = survival_label,
           ylab = "Survival probability",
           #title = varName,       
           linetype = "solid", # strata change line type by groups
           surv.median.line = "hv", # Specify median survival
           conf.int = draw.conf.int,          # Add confidence interval
           pval = TRUE,              # Add p-value
           pval.method = TRUE,
           pval.size = 10,
           ncensor.plot = FALSE,
           risk.table = TRUE,        # Add risk table
           #risk.table.col = "black",  # Risk table color by groups
           risk.table.col = "strata", 
           risk.table.height= 0.3,
           risk.table.fontsize = 10,
           risk.table.y.text = FALSE,
           risk.table.y.text.col = TRUE,
           
           #legend.labs = c("Median and above","Below median"),
           palette = c("orchid2", "dodgerblue2", "springgreen2", "saddlebrown")
       )  + theme_survminer(
         font.main = c(24, "bold", "black"),
         font.submain = c(24, "plain", "black"),
         font.caption = c(24, "plain", "orange"),
         font.x = c(24, "plain", "black"),
         font.y = c(24, "plain", "black"),
         font.tickslab = c(24, "plain", "black"),
         font.legend = c(24, "plain", "black"),
         legend = "top")
  y
}

computeHR <- function(dataTable){
  coxPH1 <- summary(coxph(Surv(Survival,Status)~(Median_and_above),data = dataTable))
  
  HR <- coxPH1$conf.int["Median_and_aboveTRUE","exp(coef)"]
  lowCI <- coxPH1$conf.int["Median_and_aboveTRUE","lower .95"]
  highCI <- coxPH1$conf.int["Median_and_aboveTRUE","upper .95"]
  HRprob <-  coxPH1$coefficients["Median_and_aboveTRUE","Pr(>|z|)"]
  logRankP <- coxPH1$sctest[["pvalue"]]
  return(c(HR,lowCI,highCI,HRprob,logRankP))
}
```

CoxPH analysis, multivariate
```{r}

## get variables with the best log rank P within their clusters, logRankP <= 0.05, 
itemList0 <- scan(file="UnivarLogRank.SelectedVars.ICI.MAPKi.txt",sep="\n",what="")
itemList0 <- c("TLS_9_GENE_SIGNATURE_Cabrita",
              "CD19",
              "CD3D",
              "IGH.diversity",
              "HALLMARK_INTERFERON_GAMMA_RESPONSE",
              "CXCR5",
              "IFNG")
```
```{r}
library(survivalAnalysis)

#inputTable <- Main_table2
#covariates_list <- covs1
#surv_variables <- survVars
#title1 <- dataset

generateForestPlot <- function(inputTable,covariates_list,surv_variables,title1){
      
      inputTable_relabel <- inputTable
      col_relabel <- colnames(inputTable_relabel)
      #col_relabel[2] <- "Dataset"
      col_relabel[3] <- surv_variables[1]
      colnames(inputTable_relabel) <- col_relabel
  
      result <- analyse_multivariate(inputTable_relabel, 
                                     vars(!!as.name(surv_variables[1]),
                                          !!as.name(surv_variables[2])),
                                     covariates = covariates_list)
      
      labels1 <- c("endpoint", "factor", "n")
      rel_width1 <- c(2,1.5,1)
      #x_breaks1 <- c(0.0, 0.2, 0.4, 0.6, 0.8, 1.0)
      
      if(surv_variables[1] == "overall_survival_day"){
                  plot1 <- forest_plot(result,
                    labels_displayed = labels1,
                    ggtheme = ggplot2::theme_bw(base_size = 8),
                    relative_widths = rel_width1,
                    title = title1)
      }else{
                  plot1 <- forest_plot(result,
                    labels_displayed = labels1,
                    ggtheme = ggplot2::theme_bw(base_size = 8),
                    relative_widths = rel_width1,
                    title = title1)
      }
      plot1
}      
```

```{r, warning=F}

## test pairwise independence of representative variables with significant univariate HR in ICI OT

#stratifyByBatch <- F
#dataset <- "ICI.OT.102622"
stratifyByBatch <- T
dataset <- "ICI.OT.BatchAdj.112222"

treatment <- c("ICI")
biopsy_timing <- c("OT")

batch_string <- ") ~"
if(stratifyByBatch){
  batch_string <- ") ~ bioproject_name+"
}

groupingColumn <- list()
groupingColumn[["MAPKi"]] <- c("progression_free_survival_day",
                               "PFS_censoring_status")
groupingColumn[["ICI"]]   <- c("overall_survival_day",
                                "OS_censoring_status")
survival_label <- list()
survival_label[["MAPKi"]] <- "PFS"
survival_label[["ICI"]] <- "OS"

panel_list <- list()
index <- 1
skipGene <- 0
treatment1 <- "ICI"
biopsy_timing1 <- "OT"
index <- 1
item_index <- 1
graph_height <- 183
graph_width <- 183

combined_table1 <- combined_table
new_colnames <- gsub("^[\\w\\_]+\\-\\-","",colnames(combined_table1),perl=T) 
colnames(combined_table1) <- new_colnames


if(stratifyByBatch){
  HR.list1 <- data.frame(Treatment.Time = character(),
                        Concordance = numeric(),
                        Batch = character(),
                        V1 = character(),
                        V2 = character(),
                        Batch.HR=numeric(), 
                        Batch.HR.P=numeric(),
                        V1.HR=numeric(), 
                        V1.HR.P=numeric(),
                        V2.HR=numeric(), 
                        V2.HR.P=numeric(),
                        stringsAsFactors = FALSE)
}else{
  HR.list1 <- data.frame(Treatment.Time = character(),
                        Concordance = numeric(),
                        V1 = character(),
                        V2 = character(),
                        V1.HR=numeric(), 
                        V1.HR.P=numeric(),
                        V2.HR=numeric(), 
                        V2.HR.P=numeric(),
                        stringsAsFactors = FALSE)
}

summaryFunction <- "maxVal"


i=4
j=5
numItem  <- length(itemList0)

for(biopsy_timing1 in biopsy_timing){
   for(treatment1 in treatment){
      for(i in 1:(numItem-1)){
        for(j in (i+1):numItem){
          
          surv_variables <- groupingColumn[[treatment1]]
          itemList1 <- c(itemList0[i],itemList0[j])
          print(itemList1)
          subset_table1 <- combined_table1 %>% 
                           filter(.,treatment==treatment1 & 
                                  biopsy_time1 == biopsy_timing1) %>% 
                                  select(all_of(c("patient_uid","bioproject_name",
                                         surv_variables,itemList1)))
          Main_table <- NA
          Main_table_empty <- 1
          #dataset <- paste(treatment1,biopsy_timing1,sep=".")
                
          for(item in itemList1){
            subset_table2 <- subset_table1 %>% 
                             select(all_of(c("patient_uid","bioproject_name",
                                         groupingColumn[[treatment1]],item)))
            colnames(subset_table2) <- c("patient_uid","bioproject_name",
                                       surv_variables,"Variable")
            subset_table2$Variable <- (as.numeric(subset_table2$Variable))
            
            ## when there are multiple tumor samples, 
            ## select the one with the highest/lowest value of Variable  
            subset_table3 <- subset_table2 %>% 
                           filter(!(is.na(Variable))) %>%
                           group_by(patient_uid,bioproject_name) %>%
                           mutate(Variable1=case_when(
                             summaryFunction=="maxVal" ~ max(Variable),
                             summaryFunction=="minVal" ~ min(Variable),
                             TRUE ~ 1)) %>%
                           ungroup %>%
                           select(all_of(c("patient_uid","bioproject_name",
                                           surv_variables,"Variable1"))) %>%
                           unique %>%
                           as.data.frame()
          
            colnames(subset_table3) <- c("patient_uid","bioproject_name",
                                       surv_variables,item)
            
            if(Main_table_empty){
              Main_table <- subset_table3
              Main_table_empty <- 0
            }
            else{
              Main_table <- Main_table %>% full_join(subset_table3)
            }
          }
      
          Main_table$bioproject_name <- factor(Main_table$bioproject_name,
                                        levels=unique(Main_table$bioproject_name))
      
          Main_table2 <- Main_table %>%
            mutate(Status = case_when(!!as.name(surv_variables[2]) == "censored" ~ 0,
                                      TRUE ~ 1)) %>%
            mutate(across(.cols=itemList1,.fns = ~ as.numeric(.)))
          
          Main_table2[[surv_variables[1]]] <- as.numeric(Main_table2[[surv_variables[1]]])
          
          coxph_wout_interaction <-  as.formula(
                                    paste0("Surv(",
                                    paste(surv_variables[1],"Status",sep=","),
                                    batch_string,
                                    paste(itemList1,collapse="+")))
          
          model_wout_interaction <- coxph(coxph_wout_interaction,data = Main_table2)
          s <- summary(model_wout_interaction) 
         
          if(stratifyByBatch){
            HR.list1[item_index,]<-c(dataset,s$concordance[1],
                                      levels(Main_table2$bioproject_name)[2],
                                      itemList1[1],
                                      itemList1[2],
                                      s$coefficients[1,c(2,5)],
                                      s$coefficients[2,c(2,5)],
                                      s$coefficients[3,c(2,5)])
          }else{
            HR.list1[item_index,]<-c(dataset,s$concordance[1],
                                      itemList1[1],
                                      itemList1[2],
                                      s$coefficients[1,c(2,5)],
                                      s$coefficients[2,c(2,5)])
          }
          
          item_index <- item_index+1
          
        }
      }
   }
}

##adjust all P values
combinedPvalList <- c(as.numeric(HR.list1$V1.HR.P),as.numeric(HR.list1$V2.HR.P))
combinedPvalListAdj <- p.adjust(combinedPvalList)
nrow1 <- nrow(HR.list1)

HR.list1$V1.HR.P.adj <- combinedPvalListAdj[c(1:nrow1)]
HR.list1$V2.HR.P.adj <- combinedPvalListAdj[c((nrow1+1):(2*nrow1))]

write.csv(HR.list1,file=paste0(dataset,".UnivarSignifVariable.Pairwise.Indep.HR.csv"))

```


```{r, warning=F}

## test interaction of representative variables with significant univariate HR in ICI OT

#stratifyByBatch <- F
#dataset <- "ICI.OT.102622"
#dataset <- "ICI.OT.BatchAdj.102622"
dataset <- "ICI.OT.BatchAdj.112222"
stratifyByBatch <- T


treatment <- c("ICI")
biopsy_timing <- c("OT")

batch_string <- ") ~"
if(stratifyByBatch){
  batch_string <- ") ~ bioproject_name+"
}

groupingColumn <- list()
groupingColumn[["MAPKi"]] <- c("progression_free_survival_day",
                               "PFS_censoring_status")
groupingColumn[["ICI"]]   <- c("overall_survival_day",
                                "OS_censoring_status")
survival_label <- list()
survival_label[["MAPKi"]] <- "PFS"
survival_label[["ICI"]] <- "OS"

panel_list <- list()
index <- 1
skipGene <- 0
treatment1 <- "ICI"
biopsy_timing1 <- "OT"
index <- 1
item_index <- 1
graph_height <- 183
graph_width <- 183

combined_table1 <- combined_table
new_colnames <- gsub("^[\\w\\_]+\\-\\-","",colnames(combined_table1),perl=T) 
colnames(combined_table1) <- new_colnames


if(stratifyByBatch){
  HR.list1 <- data.frame(Treatment.Time = character(),
                        Concordance = numeric(),
                        Batch = character(),
                        V1 = character(),
                        V2 = character(),
                        Batch.HR=numeric(), 
                        Batch.HR.P=numeric(),
                        V1.HR=numeric(), 
                        V1.HR.P=numeric(),
                        V2.HR=numeric(), 
                        V2.HR.P=numeric(),
                        Int.HR=numeric(), 
                        Int.HR.P=numeric(),
                        stringsAsFactors = FALSE)
}else{
  HR.list1 <- data.frame(Treatment.Time = character(),
                        Concordance = numeric(),
                        V1 = character(),
                        V2 = character(),
                        V1.HR=numeric(), 
                        V1.HR.P=numeric(),
                        V2.HR=numeric(), 
                        V2.HR.P=numeric(),
                        Int.HR=numeric(), 
                        Int.HR.P=numeric(),
                        stringsAsFactors = FALSE)
}

summaryFunction <- "maxVal"


i=1
j=2
numItem  <- length(itemList0)

for(biopsy_timing1 in biopsy_timing){
   for(treatment1 in treatment){
      for(i in 1:(numItem-1)){
        for(j in (i+1):numItem){
          
          surv_variables <- groupingColumn[[treatment1]]
          itemList1 <- c(itemList0[i],itemList0[j])
          
          subset_table1 <- combined_table1 %>% 
                           filter(.,treatment==treatment1 & 
                                  biopsy_time1 == biopsy_timing1) %>% 
                                  select(all_of(c("patient_uid","bioproject_name",
                                         surv_variables,itemList1)))
          Main_table <- NA
          Main_table_empty <- 1
          #dataset <- paste(treatment1,biopsy_timing1,sep=".")
                
          for(item in itemList1){
            subset_table2 <- subset_table1 %>% 
                             select(all_of(c("patient_uid","bioproject_name",
                                         groupingColumn[[treatment1]],item)))
            colnames(subset_table2) <- c("patient_uid","bioproject_name",
                                       surv_variables,"Variable")
            subset_table2$Variable <- (as.numeric(subset_table2$Variable))
            
            ## when there are multiple tumor samples, 
            ## select the one with the highest/lowest value of Variable  
            subset_table3 <- subset_table2 %>% 
                           filter(!(is.na(Variable))) %>%
                           group_by(patient_uid,bioproject_name) %>%
                           mutate(Variable1=case_when(
                             summaryFunction=="maxVal" ~ max(Variable),
                             summaryFunction=="minVal" ~ min(Variable),
                             TRUE ~ 1)) %>%
                           ungroup %>%
                           select(all_of(c("patient_uid","bioproject_name",
                                           surv_variables,"Variable1"))) %>%
                           unique %>%
                           as.data.frame()
          
            colnames(subset_table3) <- c("patient_uid","bioproject_name",
                                       surv_variables,item)
            
            if(Main_table_empty){
              Main_table <- subset_table3
              Main_table_empty <- 0
            }
            else{
              Main_table <- Main_table %>% full_join(subset_table3)
            }
          }
      
          Main_table$bioproject_name <- factor(Main_table$bioproject_name,
                                        levels=unique(Main_table$bioproject_name))
      
          Main_table2 <- Main_table %>%
            mutate(Status = case_when(!!as.name(surv_variables[2]) == "censored" ~ 0,
                                      TRUE ~ 1)) %>%
            mutate(across(.cols=itemList1,.fns = ~ as.numeric(.)))
          
          Main_table2[[surv_variables[1]]] <- as.numeric(Main_table2[[surv_variables[1]]])
          
          coxph_w_interaction <-  as.formula(
                                    paste0("Surv(",
                                    paste(surv_variables[1],"Status",sep=","),
                                    batch_string,
                                    paste(itemList1,collapse="+"),
                                    paste0("+",itemList1[1],"*",itemList1[2])))
          
          model_w_interaction <- coxph(coxph_w_interaction,data = Main_table2)
          s <- summary(model_w_interaction) 
         
          if(stratifyByBatch){
            HR.list1[item_index,]<-c(dataset,s$concordance[1],
                                      levels(Main_table2$bioproject_name)[2],
                                      itemList1[1],
                                      itemList1[2],
                                      s$coefficients[1,c(2,5)],
                                      s$coefficients[2,c(2,5)],
                                      s$coefficients[3,c(2,5)],
                                      s$coefficients[4,c(2,5)])
          }
          else{
            HR.list1[item_index,]<-c(dataset,s$concordance[1],
                                      itemList1[1],
                                      itemList1[2],
                                      s$coefficients[1,c(2,5)],
                                      s$coefficients[2,c(2,5)],
                                      s$coefficients[3,c(2,5)])
          }
          
          item_index <- item_index+1
          
        }
      }
   }
}

##adjust all P values
combinedPvalList <- c(as.numeric(HR.list1$V1.HR.P),as.numeric(HR.list1$V2.HR.P),as.numeric(HR.list1$Int.HR.P))
combinedPvalListAdj <- p.adjust(combinedPvalList)
nrow1 <- nrow(HR.list1)

HR.list1$V1.HR.P.adj <- combinedPvalListAdj[c(1:nrow1)]
HR.list1$V2.HR.P.adj <- combinedPvalListAdj[c((nrow1+1):(2*nrow1))]
HR.list1$Int.HR.P.adj <- combinedPvalListAdj[c((2*nrow1+1):(3*nrow1))]

write.csv(HR.list1,file=paste0(dataset,".UnivarSignifVariable.PairwiseInt.HR.csv"))

```





```{r,warning=F}

library(tidyverse)
library(tidytidbits)
library(survivalAnalysis)

## draw the independent and interaction model  
pairList <- list(c("IGH.diversity","IFNG"))
#pairList <- list(c("CD19",	"CD8B"))

treatment <- c("ICI","MAPKi")               
biopsy_timing <- c("PT","OT")

groupingColumn <- list()
groupingColumn[["MAPKi"]] <- c("progression_free_survival_day",
                               "PFS_censoring_status")
groupingColumn[["ICI"]]   <- c("overall_survival_day",
                                "OS_censoring_status")
survival_label <- list()
survival_label[["MAPKi"]] <- "PFS"
survival_label[["ICI"]] <- "OS"

KM_panel_list <- list()
CoxPH_panel_list <- list()

KM_index <- 1
CoxPH_index <- 1

skipGene <- 0
treatment1 <- "ICI"
biopsy_timing1 <- "OT"

graph_height <- 183
graph_width <- 183

combined_table1 <- combined_table
new_colnames <- gsub("^[\\w\\_]+\\-\\-","",colnames(combined_table1),perl=T) 
colnames(combined_table1) <- new_colnames

stratifyByBatch <- T
batchAdjStr <- "noBatchAdj"

batch_string <- ") ~"
if(stratifyByBatch){
  batch_string <- ") ~ bioproject_name+"
  batchAdjStr <- "batchAdj"
}

summaryFunction <- "maxVal"
i=1


for(i in 1:length(pairList)){
  KM_panel_list <- list()
  CoxPH_panel_list <- list()
  
  KM_index <- 1
  CoxPH_index <- 1
  
  for(biopsy_timing1 in biopsy_timing){
    for(treatment1 in treatment){
      itemList1 <- pairList[[i]]
      surv_variables <- groupingColumn[[treatment1]]
      subset_table1 <- combined_table1 %>% 
                       filter(.,treatment==treatment1 & 
                              biopsy_time1 == biopsy_timing1) %>% 
                              select(all_of(c("patient_uid","bioproject_name",
                                     surv_variables,itemList1)))
      Main_table <- NA
      Main_table_empty <- 1
      dataset <- paste(treatment1,biopsy_timing1,sep=".")
            
      for(item in itemList1){
        subset_table2 <- subset_table1 %>% 
                         select(all_of(c("patient_uid","bioproject_name",
                                     groupingColumn[[treatment1]],item)))
        colnames(subset_table2) <- c("patient_uid","bioproject_name",
                                   surv_variables,"Variable")
        subset_table2$Variable <- (as.numeric(subset_table2$Variable))
        
        ## when there are multiple tumor samples, 
        ## select the one with the highest/lowest value of Variable  
        subset_table3 <- subset_table2 %>% 
                       filter(!(is.na(Variable))) %>%
                       group_by(patient_uid,bioproject_name) %>%
                       mutate(Variable1=case_when(
                         summaryFunction=="maxVal" ~ max(Variable),
                         summaryFunction=="minVal" ~ min(Variable),
                         TRUE ~ 1)) %>%
                       ungroup %>%
                       select(all_of(c("patient_uid","bioproject_name",
                                       surv_variables,"Variable1"))) %>%
                       unique %>%
                       as.data.frame()
      
        colnames(subset_table3) <- c("patient_uid","bioproject_name",
                                   surv_variables,item)
        
        if(Main_table_empty){
          Main_table <- subset_table3
          Main_table_empty <- 0
        }
        else{
          Main_table <- Main_table %>% full_join(subset_table3)
        }
      }
      
      int_item <- "Interaction"
      Main_table[[int_item]] = 
              Main_table[[itemList1[1]]]*Main_table[[itemList1[2]]]
      Main_table$bioproject_name <- factor(Main_table$bioproject_name,
                                        levels=unique(Main_table$bioproject_name))
      
      new_vars<-sapply(itemList1, function (x) paste0(x,"_MedianUp"))
      
      Main_table2 <- Main_table %>%
        mutate(Status = case_when(!!as.name(surv_variables[2]) == "censored" ~ 0,
                                  TRUE ~ 1)) %>%
        mutate(across(.cols=itemList1,.fns = ~ as.numeric(.))) %>%
        mutate(across(.cols=itemList1,.fns = list(MedianUp = ~.x>=median(.x,na.rm = T))))
     
      Main_table2[[surv_variables[1]]] <- as.numeric(Main_table2[[surv_variables[1]]])
      
      Main_table2[["Interaction_strata"]] = 
              paste0(as.character(Main_table2[[paste0(itemList1[1],"_MedianUp")]]),
                 "_",as.character(Main_table2[[paste0(itemList1[2],"_MedianUp")]]))
      
      Main_table2[[ncol(Main_table2)]] = 
                      factor(Main_table2[[ncol(Main_table2)]], 
                                    levels=c("TRUE_TRUE","TRUE_FALSE",
                                             "FALSE_TRUE","FALSE_FALSE"))

      #subset1 <- Main_table2[Main_table2$IGH.diversity_MedianUp=="TRUE",]
      #table(subset1$bioproject_name)
      #table(Main_table2$bioproject_name)
      

      
      ##independent model
      if(stratifyByBatch){
        covs1 <- vars(bioproject_name,!!as.name(itemList1[1]),!!as.name(itemList1[2]))
      }else{
        covs1 <- vars(!!as.name(itemList1[1]),!!as.name(itemList1[2]))
      }
      survVars <- c(survival_label[[treatment1]],"Status")
      
      coxph_continuous <-  as.formula(paste0("Surv(",
                          paste(surv_variables[1],"Status",sep=","),
                          batch_string,
                          paste(itemList1,collapse="+")))
      model_continuous <- coxph(coxph_continuous,data = Main_table2)
      test.model_continuous <- cox.zph(model_continuous)
      
      # CoxPH_panel_list[[CoxPH_index]] <- ggforest(model_continuous, 
      #                                     main=paste(dataset,"noInt",sep="_"),
      #                                             fontsize = 1,
      #                                             noDigits = 1)
      
      CoxPH_panel_list[[CoxPH_index]] <- generateForestPlot(Main_table2,
                                                            covs1,
                                                            survVars,
                                                            dataset)
      CoxPH_index <- CoxPH_index+1
      
      
      
      ##interaction model
      
      coxph_w_interaction <-  as.formula(paste0("Surv(",
                    paste(surv_variables[1],"Status",sep=","),
                    batch_string,
                    paste(itemList1,collapse="+"),
                    paste0("+Interaction")))
      
      model_w_interaction <- coxph(coxph_w_interaction,data = Main_table2)
      s <- summary(model_w_interaction) 
      test.model_w_interaction <- cox.zph(model_w_interaction)
      
      # CoxPH_panel_list[[CoxPH_index]] <- ggforest(model_w_interaction,
      #                                             main=paste(dataset,"Int",sep="_"),
      #                                             fontsize = 1,
      #                                             noDigits = 1)
      
      if(stratifyByBatch){
        covs1 <- vars(bioproject_name,!!as.name(itemList1[1]),
                      !!as.name(itemList1[2]),!!as.name("Interaction"))
      }else{
        covs1 <- vars(!!as.name(itemList1[1]),!!as.name(itemList1[2]),
                      !!as.name("Interaction"))
      }
      
      CoxPH_panel_list[[CoxPH_index]] <- generateForestPlot(Main_table2,
                                                            covs1,
                                                            survVars,
                                                            dataset)
      CoxPH_index <- CoxPH_index+1

      
      #Draw the KM for the two interacting variables
      Main_table3 <- Main_table2[c(surv_variables[1],"Status","Interaction_strata")]
      colnames(Main_table3) <- c("Survival","Status","Value")
      
      KM_panel_list[[KM_index]] <- generateKaplanMeier(Main_table3,
                                             "Interaction_strata",
                                             survival_label[[treatment1]],
                                             draw.conf.int = F)
      KM_index <- KM_index+1
      
    }
  }
  
  ncol1 = 2
  nrow1 = 2
  width1 <- ncol1*graph_width*1.3
  height1 <- nrow1*graph_height*1.3
  plot1 <- arrange_ggsurvplots(
            KM_panel_list,
            title = paste(itemList1[1],itemList1[2],sep="_"),
            ncol = ncol1,
            nrow = nrow1,
            print = FALSE
  )
 
  ggsave(filename = paste("KMplot",batchAdjStr,paste(itemList1[1],itemList1[2],sep="_"),"pdf",sep="."),
           plot = plot1,
           device = "pdf",
           width = width1,
           height = height1,
           units = "mm")
  
  ncol1 = 1
  nrow1 = 1
  width1 <- ncol1*graph_width*2.3
  height1 <- nrow1*graph_height*1.3

  library(gridExtra)
  plot1 <- marrangeGrob(
            CoxPH_panel_list,
            ncol = ncol1,
            nrow = nrow1,
            top = NULL)
  
  width1 <- 196
  height1 <- 35
  ggsave(filename = paste("CoxPHplot",batchAdjStr,paste(itemList1[1],itemList1[2],sep="_"),"pdf",sep="."),
           plot = plot1,
           device = "pdf",
           width = width1,
           height = height1,
           units = "mm")
  
}

```


Draw all variables. independent
```{r,warning=F}

library(tidyverse)
library(tidytidbits)
library(survivalAnalysis)

## draw the independent model of all signif univariate
pairList <- list(itemList0)

treatment <- c("ICI","MAPKi")               
biopsy_timing <- c("PT","OT")

groupingColumn <- list()
groupingColumn[["MAPKi"]] <- c("progression_free_survival_day",
                               "PFS_censoring_status")
groupingColumn[["ICI"]]   <- c("overall_survival_day",
                                "OS_censoring_status")
survival_label <- list()
survival_label[["MAPKi"]] <- "PFS"
survival_label[["ICI"]] <- "OS"

KM_panel_list <- list()
CoxPH_panel_list <- list()

KM_index <- 1
CoxPH_index <- 1

skipGene <- 0
treatment1 <- "ICI"
biopsy_timing1 <- "OT"

graph_height <- 183
graph_width <- 183

combined_table1 <- combined_table
new_colnames <- gsub("^[\\w\\_]+\\-\\-","",colnames(combined_table1),perl=T) 
colnames(combined_table1) <- new_colnames

stratifyByBatch <- T
batchAdjStr <- "noBatchAdj"

batch_string <- ") ~"
if(stratifyByBatch){
  batch_string <- ") ~ bioproject_name+"
  batchAdjStr <- "batchAdj"
}

summaryFunction <- "maxVal"
i=1


for(i in 1:length(pairList)){
  KM_panel_list <- list()
  CoxPH_panel_list <- list()
  
  KM_index <- 1
  CoxPH_index <- 1
  
  for(biopsy_timing1 in biopsy_timing){
    for(treatment1 in treatment){
      itemList1 <- pairList[[i]]
      surv_variables <- groupingColumn[[treatment1]]
      subset_table1 <- combined_table1 %>% 
                       filter(.,treatment==treatment1 & 
                              biopsy_time1 == biopsy_timing1) %>% 
                              select(all_of(c("patient_uid","bioproject_name",
                                     surv_variables,itemList1)))
      Main_table <- NA
      Main_table_empty <- 1
      dataset <- paste(treatment1,biopsy_timing1,sep=".")
            
      for(item in itemList1){
        subset_table2 <- subset_table1 %>% 
                         select(all_of(c("patient_uid","bioproject_name",
                                     groupingColumn[[treatment1]],item)))
        colnames(subset_table2) <- c("patient_uid","bioproject_name",
                                   surv_variables,"Variable")
        subset_table2$Variable <- (as.numeric(subset_table2$Variable))
        
        ## when there are multiple tumor samples, 
        ## select the one with the highest/lowest value of Variable  
        subset_table3 <- subset_table2 %>% 
                       filter(!(is.na(Variable))) %>%
                       group_by(patient_uid,bioproject_name) %>%
                       mutate(Variable1=case_when(
                         summaryFunction=="maxVal" ~ max(Variable),
                         summaryFunction=="minVal" ~ min(Variable),
                         TRUE ~ 1)) %>%
                       ungroup %>%
                       select(all_of(c("patient_uid","bioproject_name",
                                       surv_variables,"Variable1"))) %>%
                       unique %>%
                       as.data.frame()
      
        colnames(subset_table3) <- c("patient_uid","bioproject_name",
                                   surv_variables,item)
        
        if(Main_table_empty){
          Main_table <- subset_table3
          Main_table_empty <- 0
        }
        else{
          Main_table <- Main_table %>% full_join(subset_table3)
        }
      }
      
      Main_table$bioproject_name <- factor(Main_table$bioproject_name,
                                        levels=unique(Main_table$bioproject_name))
      
      new_vars<-sapply(itemList1, function (x) paste0(x,"_MedianUp"))
      
      Main_table2 <- Main_table %>%
        mutate(Status = case_when(!!as.name(surv_variables[2]) == "censored" ~ 0,
                                  TRUE ~ 1)) %>%
        mutate(across(.cols=itemList1,.fns = ~ as.numeric(.))) %>%
        mutate(across(.cols=itemList1,.fns = list(MedianUp = ~.x>=median(.x,na.rm = T))))
     
      Main_table2[[surv_variables[1]]] <- as.numeric(Main_table2[[surv_variables[1]]])
    
      ##independent model
      if(stratifyByBatch){
        covs1 <- vars(bioproject_name,!!as.name(itemList1[1]),!!as.name(itemList1[2]))
      }else{
        covs1 <- vars(!!as.name(itemList1[1]),!!as.name(itemList1[2]))
      }
      survVars <- c(survival_label[[treatment1]],"Status")
      
      coxph_continuous <-  as.formula(paste0("Surv(",
                          paste(surv_variables[1],"Status",sep=","),
                          batch_string,
                          paste(itemList1,collapse="+")))
      model_continuous <- coxph(coxph_continuous,data = Main_table2)
      test.model_continuous <- cox.zph(model_continuous)
      
      # CoxPH_panel_list[[CoxPH_index]] <- ggforest(model_continuous, 
      #                                     main=paste(dataset,"noInt",sep="_"),
      #                                             fontsize = 1,
      #                                             noDigits = 1)
      
      CoxPH_panel_list[[CoxPH_index]] <- generateForestPlot(Main_table2,
                                                            covs1,
                                                            survVars,
                                                            dataset)
      CoxPH_index <- CoxPH_index+1
      
    }
  }
  
  ncol1 = 1
  nrow1 = 1
  width1 <- ncol1*graph_width*2.3
  height1 <- nrow1*graph_height*1.3

  library(gridExtra)
  plot1 <- marrangeGrob(
            CoxPH_panel_list,
            ncol = ncol1,
            nrow = nrow1,
            top = NULL)
  
  width1 <- 196
  height1 <- 35
  ggsave(filename = paste("CoxPHplot",batchAdjStr,paste(itemList1[1],itemList1[2],sep="_"),"pdf",sep="."),
           plot = plot1,
           device = "pdf",
           width = width1,
           height = height1,
           units = "mm")
  
}

```


Not used
```{r}

itemList0 <- scan(file="UnivarLogRank.SelectedVars.ICI.MAPKi.txt",sep="\n",what="")
stratifyByBatch <- F

batch_string <- ") ~"
if(stratifyByBatch){
  batch_string <- ") ~ bioproject_name+"
  batchAdjStr <- "batchAdj"
}
groupingColumn <- list()
groupingColumn[["MAPKi"]] <- c("progression_free_survival_day",
                               "PFS_censoring_status")
groupingColumn[["ICI"]]   <- c("overall_survival_day",
                                "OS_censoring_status")

survival_label <- list()
survival_label[["MAPKi"]] <- "PFS"
survival_label[["ICI"]] <- "OS"

panel_list <- list()
index <- 1
skipGene <- 0
treatment1 <- "ICI"
biopsy_timing1 <- "OT"


surv_variables <- groupingColumn[[treatment1]]
itemList1 <- itemList0

subset_table1 <- combined_table1 %>% 
                 filter(.,treatment==treatment1 & 
                        biopsy_time1 == biopsy_timing1) %>% 
                        select(all_of(c("patient_uid","bioproject_name",
                               surv_variables,itemList1)))
Main_table <- NA
Main_table_empty <- 1
#dataset <- paste(treatment1,biopsy_timing1,sep=".")
      
for(item in itemList1){
  subset_table2 <- subset_table1 %>% 
                   select(all_of(c("patient_uid","bioproject_name",
                               groupingColumn[[treatment1]],item)))
  colnames(subset_table2) <- c("patient_uid","bioproject_name",
                             surv_variables,"Variable")
  subset_table2$Variable <- (as.numeric(subset_table2$Variable))
  
  ## when there are multiple tumor samples, 
  ## select the one with the highest/lowest value of Variable  
  subset_table3 <- subset_table2 %>% 
                 filter(!(is.na(Variable))) %>%
                 group_by(patient_uid,bioproject_name) %>%
                 mutate(Variable1=case_when(
                   summaryFunction=="maxVal" ~ max(Variable),
                   summaryFunction=="minVal" ~ min(Variable),
                   TRUE ~ 1)) %>%
                 ungroup %>%
                 select(all_of(c("patient_uid","bioproject_name",
                                 surv_variables,"Variable1"))) %>%
                 unique %>%
                 as.data.frame()

  colnames(subset_table3) <- c("patient_uid","bioproject_name",
                             surv_variables,item)
  
  if(Main_table_empty){
    Main_table <- subset_table3
    Main_table_empty <- 0
  }
  else{
    Main_table <- Main_table %>% full_join(subset_table3)
  }
}

Main_table$bioproject_name <- factor(Main_table$bioproject_name,
                              levels=unique(Main_table$bioproject_name))

Main_table2 <- Main_table %>%
  mutate(Status = case_when(!!as.name(surv_variables[2]) == "censored" ~ 0,
                            TRUE ~ 1)) %>%
  mutate(across(.cols=itemList1,.fns = ~ as.numeric(.)))

Main_table2[[surv_variables[1]]] <- as.numeric(Main_table2[[surv_variables[1]]])
          


coxph_wout_interaction <-  as.formula(
                                    paste0("Surv(",
                                    paste(surv_variables[1],"Status",sep=","),
                                    batch_string,
                                    paste(itemList0,collapse="+")))
          
model_wout_interaction <- coxph(coxph_wout_interaction,data = Main_table2)
s <- summary(model_wout_interaction) 
s
```
